finalize refactor of nodepools related utils
