+++ services/terraformer/server/domain/utils/cluster-builder/cluster_builder.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"iter"
 	"net"
 	"os"
 	"path/filepath"
@@ -130,7 +129,7 @@ func (c ClusterBuilder) CreateNodepools() error {
 			continue
 		}
 
-		f := hash.Digest128(filepath.Join(np.Provider.SpecName, templates.ExtractTargetPath(np.Provider.Templates)))
+		f := hash.Digest128(filepath.Join(np.Provider.SpecName, np.Provider.Templates.MustExtractTargetPath()))
 		k := fmt.Sprintf("%s_%s_%s", nodepool.Name, np.Provider.SpecName, hex.EncodeToString(f))
 
 		output, err := terraform.Output(k)
@@ -230,10 +229,10 @@ func (c *ClusterBuilder) generateFiles(clusterID, clusterDir string) error {
 		return fmt.Errorf("error while generating provider templates: %w", err)
 	}
 
-	for info, pools := range GroupByProvider(clusterInfo.NodePools) {
+	for info, pools := range nodepools.ByProvider(clusterInfo.NodePools) {
 		templatesDownloadDir := filepath.Join(TemplatesRootDir, clusterID, info.SpecName)
 
-		for path, pools := range GroupByTemplates(pools) {
+		for path, pools := range nodepools.ByTemplates(pools) {
 			p := pools[0].GetDynamicNodePool().GetProvider()
 
 			if err := templates.DownloadProvider(templatesDownloadDir, p); err != nil {
@@ -421,14 +420,14 @@ func (c *ClusterBuilder) generateProviderTemplates(current, desired *spec.Cluste
 		slices.Values(desired.GetNodePools()),
 	)
 
-	for info, pools := range GroupByProvider(nps) {
+	for info, pools := range nodepools.ByProvider(nps) {
 		if err := fileutils.CreateKey(info.Creds, directory, info.SpecName); err != nil {
 			return fmt.Errorf("error creating provider credential key file for provider %s in %s : %w", info.SpecName, directory, err)
 		}
 
 		templatesDownloadDir := filepath.Join(TemplatesRootDir, clusterID, info.SpecName)
 
-		for path, pools := range GroupByTemplates(pools) {
+		for path, pools := range nodepools.ByTemplates(pools) {
 			p := pools[0].GetDynamicNodePool().GetProvider()
 			if err := templates.DownloadProvider(templatesDownloadDir, p); err != nil {
 				msg := fmt.Sprintf("cluster %q failed to download template repository", clusterID)
@@ -457,57 +456,4 @@ func (c *ClusterBuilder) generateProviderTemplates(current, desired *spec.Cluste
 	}
 
 	return nil
-}
-
-type ProviderTemplateGroup struct {
-	CloudProvider string
-	SpecName      string
-	Creds         string
-}
-
-func GroupByProvider(nps []*spec.NodePool) iter.Seq2[ProviderTemplateGroup, []*spec.NodePool] {
-	m := make(map[ProviderTemplateGroup][]*spec.NodePool)
-
-	for _, nodepool := range nps {
-		np, ok := nodepool.Type.(*spec.NodePool_DynamicNodePool)
-		if !ok {
-			continue
-		}
-		k := ProviderTemplateGroup{
-			CloudProvider: np.DynamicNodePool.Provider.CloudProviderName,
-			SpecName:      np.DynamicNodePool.Provider.SpecName,
-			Creds:         np.DynamicNodePool.Provider.Credentials(),
-		}
-		m[k] = append(m[k], nodepool)
-	}
-
-	return func(yield func(ProviderTemplateGroup, []*spec.NodePool) bool) {
-		for k, v := range m {
-			if !yield(k, v) {
-				return
-			}
-		}
-	}
-}
-
-func GroupByTemplates(nps []*spec.NodePool) iter.Seq2[string, []*spec.NodePool] {
-	m := make(map[string][]*spec.NodePool)
-
-	for _, nodepool := range nps {
-		np, ok := nodepool.Type.(*spec.NodePool_DynamicNodePool)
-		if !ok {
-			continue
-		}
-
-		p := templates.ExtractTargetPath(np.DynamicNodePool.Provider.Templates)
-		m[p] = append(m[p], nodepool)
-	}
-
-	return func(yield func(string, []*spec.NodePool) bool) {
-		for k, v := range m {
-			if !yield(k, v) {
-				return
-			}
-		}
-	}
 }
